Fix: complete insta_taged domain
